requirements.txt
git+https://github.com/dolfies/discord.py-self.git
python-dotenv==1.0.0
